aplicando autorizacion con policy y restriccion de vista de botones que no deberia hacer alguien que no sea admin

+++ app/Http/Controllers/Admin/GrupoController.php
@@ -18,7 +18,7 @@ class GrupoController extends Controller
     */
     public function __construct(MoodleServiceProvider $moodleService)
     {
-        $this->authorizeResource(Grupo::class, 'grupo');
+        //lo hago manual la autorizacion de recursos
         $this->moodleService = $moodleService;
     }
 
@@ -27,7 +27,13 @@ class GrupoController extends Controller
      */
     public function index(Edicion $edicion)
     {
-        $grupos = $edicion->grupos()->get();
+        //le estoy indicando que esa edicion esta en grupo porque laravel se piensa que ese metodo esta en edicionPolicy porque la instancia es de edicion
+        $this->authorize('soloVerGruposEdicion', [Grupo::class, $edicion]);
+        if (Auth::user()->isAdmin()) { //parecido a EdicionController el bloque
+            $grupos = $edicion->grupos()->get();
+        } else {
+            $grupos = $edicion->grupos()->where('tutor', Auth::id())->get();
+        }
         return view('admin.grupos.index', ['grupos' => $grupos, 'edicion' => $edicion]);
     }
 
@@ -36,6 +42,7 @@ class GrupoController extends Controller
      */
     public function show(Grupo $grupo)
     {
+        $this->authorize('view', $grupo);
         return view('admin.grupos.show', compact('grupo'));
     }
 
@@ -70,6 +77,7 @@ class GrupoController extends Controller
 
     public function edit(Grupo $grupo)
     {
+        $this->authorize('update', $grupo);
         $edicion = $grupo->edicion;
         return view('admin.grupos.edit', compact('grupo', 'edicion'));
     }
@@ -78,6 +86,7 @@ class GrupoController extends Controller
      */
 
     public function update(Request $request, Grupo $grupo){
+        $this->authorize('update', $grupo);
         $edicion = $grupo->edicion;
         $request->validate([
             'nombre' => 'required|max:100',
@@ -95,6 +104,7 @@ class GrupoController extends Controller
 
     public function destroy(Grupo $grupo)
     {
+        $this->authorize('delete', $grupo);
         $edicion = $grupo->edicion;
         $grupo->delete();
         return redirect()->route('ediciones.grupos.index', ['edicion' => $edicion])->with('success', 'Grupo eliminado correctamente.');
